application-generator/src/content/python/modules/user/requirements.txt
alembic==1.8.0
aniso8601==9.0.1
apifairy==1.1.0
apispec==5.2.2
blinker==1.4
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
dominate==2.6.0
Flask==2.1.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-DotEnv==0.1.2
Flask-HTTPAuth==4.7.0
Flask-Login==0.6.1
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==2.6.0
Flask-RESTful==0.3.9
Flask-Seeder==1.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.0
mysql-connector-python==8.0.29
packaging==21.3
protobuf==4.21.2
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
visitor==0.1.3
webargs==8.2.0
Werkzeug==2.1.2
WTForms==3.0.1